fixed gotcha in ensuring esusu group got a hash id on creation
In the signals.py file, the instance was not saved to database after
hash id was generated for and attached to it.
So, while the instance itself had a hash id, what had been persisted
to the database remained without.

This escaped tests as the tests were operating on the in-memory
instance